% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{get_ld}
\alias{get_ld}
\title{Title}
\usage{
get_ld(haploA, haploB, stat = "r2")
}
\arguments{
\item{haploA}{Vector of haplotypes for the first marker, coded 0/1.}

\item{haploB}{Vector of haplotypes for the second marker, coded 0/1.}

\item{stat}{What LD statistic to return ("D" for D or "r2" for r squared).}
}
\value{
Statistic (D or r squared).
}
\description{
Title
}
