% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{get_ld_matrix}
\alias{get_ld_matrix}
\title{Title}
\usage{
get_ld_matrix(haplo, stat = "r2")
}
\arguments{
\item{haplo}{Matrix of haplotypes, coded 0/1. Columns should be markers and
rows should be observed chromosomes.}

\item{stat}{What LD statistic to return ("D" for D or "r2" for r squared).}
}
\value{
A matrix of LD estimates.
}
\description{
Title
}
