% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_populations}
\alias{simulate_populations}
\title{Simulate two related populations}
\usage{
simulate_populations()
}
\value{
A list containing two matrices of haplotypes.
}
\description{
Simulates haplotypes from two populations that split 100 generations
ago, with 100 individuals from each population and 500 markers.
}
